+++ CHANGELOG.md
@@ -34,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#1893](https://github.com/open-telemetry/opentelemetry-python/pull/1893))
 
 ### Added
+- Give OTLPHandler the ability to process attributes
+  ([#1952](https://github.com/open-telemetry/opentelemetry-python/pull/1952))
 - Add global LogEmitterProvider and convenience function get_log_emitter
   ([#1901](https://github.com/open-telemetry/opentelemetry-python/pull/1901))
 - Add OTLPHandler for standard library logging module